LaTeXify-root/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "latexify"
version = "0.1.0"
description = "LaTeX document processing and synthesis toolkit"
readme = "README.md"
authors = [{name = "LaTeXify Team"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pypdf>=3.9.0",
    "typer>=0.12.5",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
faiss = ["faiss-cpu>=1.7.4"]
ocr = ["easyocr>=1.7.1", "pytesseract>=0.3.10", "opencv-python>=4.8.0"]
llama = ["llama-cpp-python>=0.2.90"]
hf-runner = [
    "transformers>=4.37.2",
    "accelerate>=0.28.0",
    "pillow>=10.3.0",
    "torch>=2.1.0",
]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.6.3",
    "pre-commit>=3.7.0",
    "aiohttp>=3.9.5",
    "chktex>=1.7.9 ; python_version < '0'",
]

[project.scripts]
latexify = "latexify.cli:run"
latexify-aggregate = "latexify.pipeline.aggregator:main"
latexify-planner = "latexify.pipeline.planner_scaffold:main"
latexify-retrieval-bundle = "latexify.pipeline.retrieval_bundle:main"
latexify-query-index = "latexify.kb.query_index:main"
latexify-build-latex-kb = "latexify.kb.build_latex_kb:main"
latexify-build-index-bgem3 = "latexify.kb.build_index_bgem3:main"
latexify-build-index-compat = "latexify.kb.build_index_compat:main"
latexify-ensure-kb-alias = "latexify.kb.ensure_kb_alias:main"
latexify-class-probe = "latexify.agents.class_probe:main"
latexify-synth-latex = "latexify.pipeline.synth_latex:main"
latexify-phase2-run-task = "latexify.pipeline.phase2_run_task:main"
latexify-setup-classes = "latexify.utils.setup_classes:main"
latexify-latex-deps-check = "latexify.utils.latex_deps_check:main"
latexify-ingest-pdf = "latexify.ingestion.ingest_pdf:main"
latexify-judgecouncil = "latexify.pipeline.judge_model:main"
latexify-judge-council = "latexify.pipeline.judge_model:main"
latexify-layout-planner = "latexify.pipeline.layout_planner:main"
latexify-tex-assembler = "latexify.pipeline.tex_assembler:main"
latexify-langchain-orchestrator = "latexify.pipeline.langchain_orchestrator:main"

[tool.setuptools.packages.find]
include = ["latexify*"]
